moved cosmosdb scalemonitor tests to separate test class
