+++ test/WebJobs.Extensions.CosmosDB.Tests/Trigger/CosmosDBListenerTests.cs
@@ -74,172 +74,6 @@ namespace Microsoft.Azure.WebJobs.Extensions.CosmosDB.Tests.Trigger
         }
 
         [Fact]
-        public void ScaleMonitorDescriptor_ReturnsExpectedValue()
-        {
-            Assert.Equal($"{_functionId}-cosmosdbtrigger-{DatabaseName}-{ContainerName}".ToLower(), _listener.GetMonitor().Descriptor.Id);
-        }
-
-        [Fact]
-        public void GetScaleStatus_NoMetrics_ReturnsVote_None()
-        {
-            var context = new ScaleStatusContext<CosmosDBTriggerMetrics>
-            {
-                WorkerCount = 1
-            };
-
-            var status = ((CosmosDBScaleMonitor)_listener.GetMonitor()).GetScaleStatus(context);
-            Assert.Equal(ScaleVote.None, status.Vote);
-
-            // verify the non-generic implementation works properly
-            status = _listener.GetMonitor().GetScaleStatus(context);
-            Assert.Equal(ScaleVote.None, status.Vote);
-        }
-
-        [Fact]
-        public void GetScaleStatus_InstancesPerPartitionThresholdExceeded_ReturnsVote_ScaleIn()
-        {
-            var context = new ScaleStatusContext<CosmosDBTriggerMetrics>
-            {
-                WorkerCount = 2
-            };
-            var timestamp = DateTime.UtcNow;
-            context.Metrics = new List<CosmosDBTriggerMetrics>
-            {
-                new CosmosDBTriggerMetrics { RemainingWork = 2500, PartitionCount = 1, Timestamp = timestamp.AddSeconds(15) },
-                new CosmosDBTriggerMetrics { RemainingWork = 2505, PartitionCount = 1, Timestamp = timestamp.AddSeconds(15) },
-                new CosmosDBTriggerMetrics { RemainingWork = 2612, PartitionCount = 1, Timestamp = timestamp.AddSeconds(15) },
-                new CosmosDBTriggerMetrics { RemainingWork = 2700, PartitionCount = 1, Timestamp = timestamp.AddSeconds(15) },
-                new CosmosDBTriggerMetrics { RemainingWork = 2810, PartitionCount = 1, Timestamp = timestamp.AddSeconds(15) },
-                new CosmosDBTriggerMetrics { RemainingWork = 2900, PartitionCount = 1, Timestamp = timestamp.AddSeconds(15) },
-            };
-
-            var status = ((CosmosDBScaleMonitor)_listener.GetMonitor()).GetScaleStatus(context);
-            Assert.Equal(ScaleVote.ScaleIn, status.Vote);
-
-            var logs = _loggerProvider.GetAllLogMessages().ToArray();
-            var log = logs[0];
-            Assert.Equal(Microsoft.Extensions.Logging.LogLevel.Information, log.Level);
-            Assert.Equal("WorkerCount (2) > PartitionCount (1).", log.FormattedMessage);
-            log = logs[1];
-            Assert.Equal(Microsoft.Extensions.Logging.LogLevel.Information, log.Level);
-            Assert.Equal($"Number of instances (2) is too high relative to number of partitions for container ({ContainerName}, 1).", log.FormattedMessage);
-        }
-
-        [Fact]
-        public void GetScaleStatus_MessagesPerWorkerThresholdExceeded_ReturnsVote_ScaleOut()
-        {
-            var context = new ScaleStatusContext<CosmosDBTriggerMetrics>
-            {
-                WorkerCount = 1
-            };
-            var timestamp = DateTime.UtcNow;
-            context.Metrics = new List<CosmosDBTriggerMetrics>
-            {
-                new CosmosDBTriggerMetrics { RemainingWork = 2500, PartitionCount = 0, Timestamp = timestamp.AddSeconds(15) },
-                new CosmosDBTriggerMetrics { RemainingWork = 2505, PartitionCount = 0, Timestamp = timestamp.AddSeconds(15) },
-                new CosmosDBTriggerMetrics { RemainingWork = 2612, PartitionCount = 0, Timestamp = timestamp.AddSeconds(15) },
-                new CosmosDBTriggerMetrics { RemainingWork = 2700, PartitionCount = 0, Timestamp = timestamp.AddSeconds(15) },
-                new CosmosDBTriggerMetrics { RemainingWork = 2810, PartitionCount = 0, Timestamp = timestamp.AddSeconds(15) },
-                new CosmosDBTriggerMetrics { RemainingWork = 2900, PartitionCount = 0, Timestamp = timestamp.AddSeconds(15) },
-            };
-
-            var status = ((CosmosDBScaleMonitor)_listener.GetMonitor()).GetScaleStatus(context);
-            Assert.Equal(ScaleVote.ScaleOut, status.Vote);
-
-            var logs = _loggerProvider.GetAllLogMessages().ToArray();
-            var log = logs[0];
-            Assert.Equal(Microsoft.Extensions.Logging.LogLevel.Information, log.Level);
-            Assert.Equal("RemainingWork (2900) > WorkerCount (1) * 1,000.", log.FormattedMessage);
-            log = logs[1];
-            Assert.Equal(Microsoft.Extensions.Logging.LogLevel.Information, log.Level);
-            Assert.Equal($"Remaining work for container ({ContainerName}, 2900) is too high relative to the number of instances (1).", log.FormattedMessage);
-        }
-
-        [Fact]
-        public void GetScaleStatus_ConsistentRemainingWork_ReturnsVote_ScaleOut()
-        {
-            var context = new ScaleStatusContext<CosmosDBTriggerMetrics>
-            {
-                WorkerCount = 1
-            };
-            var timestamp = DateTime.UtcNow;
-            context.Metrics = new List<CosmosDBTriggerMetrics>
-            {
-                new CosmosDBTriggerMetrics { RemainingWork = 10, PartitionCount = 2, Timestamp = timestamp.AddSeconds(15) },
-                new CosmosDBTriggerMetrics { RemainingWork = 10, PartitionCount = 2, Timestamp = timestamp.AddSeconds(15) },
-                new CosmosDBTriggerMetrics { RemainingWork = 10, PartitionCount = 2, Timestamp = timestamp.AddSeconds(15) },
-                new CosmosDBTriggerMetrics { RemainingWork = 10, PartitionCount = 2, Timestamp = timestamp.AddSeconds(15) },
-                new CosmosDBTriggerMetrics { RemainingWork = 10, PartitionCount = 2, Timestamp = timestamp.AddSeconds(15) },
-                new CosmosDBTriggerMetrics { RemainingWork = 10, PartitionCount = 2, Timestamp = timestamp.AddSeconds(15) },
-            };
-
-            var status = ((CosmosDBScaleMonitor)_listener.GetMonitor()).GetScaleStatus(context);
-            Assert.Equal(ScaleVote.ScaleOut, status.Vote);
-
-            var logs = _loggerProvider.GetAllLogMessages().ToArray();
-            var log = logs[0];
-            Assert.Equal(Microsoft.Extensions.Logging.LogLevel.Information, log.Level);
-            Assert.Equal($"CosmosDB container '{ContainerName}' has documents waiting to be processed.", log.FormattedMessage);
-            log = logs[1];
-            Assert.Equal(Microsoft.Extensions.Logging.LogLevel.Information, log.Level);
-            Assert.Equal("There are 1 instances relative to 2 partitions.", log.FormattedMessage);
-        }
-
-        [Fact]
-        public void GetScaleStatus_RemainingWorkIncreasing_ReturnsVote_ScaleOut()
-        {
-            var context = new ScaleStatusContext<CosmosDBTriggerMetrics>
-            {
-                WorkerCount = 1
-            };
-            var timestamp = DateTime.UtcNow;
-            context.Metrics = new List<CosmosDBTriggerMetrics>
-            {
-                new CosmosDBTriggerMetrics { RemainingWork = 10, PartitionCount = 0, Timestamp = timestamp.AddSeconds(15) },
-                new CosmosDBTriggerMetrics { RemainingWork = 20, PartitionCount = 0, Timestamp = timestamp.AddSeconds(15) },
-                new CosmosDBTriggerMetrics { RemainingWork = 40, PartitionCount = 0, Timestamp = timestamp.AddSeconds(15) },
-                new CosmosDBTriggerMetrics { RemainingWork = 80, PartitionCount = 0, Timestamp = timestamp.AddSeconds(15) },
-                new CosmosDBTriggerMetrics { RemainingWork = 100, PartitionCount = 0, Timestamp = timestamp.AddSeconds(15) },
-                new CosmosDBTriggerMetrics { RemainingWork = 150, PartitionCount = 0, Timestamp = timestamp.AddSeconds(15) },
-            };
-
-            var status = ((CosmosDBScaleMonitor)_listener.GetMonitor()).GetScaleStatus(context);
-            Assert.Equal(ScaleVote.ScaleOut, status.Vote);
-
-            var logs = _loggerProvider.GetAllLogMessages().ToArray();
-            var log = logs[0];
-            Assert.Equal(Microsoft.Extensions.Logging.LogLevel.Information, log.Level);
-            Assert.Equal($"Remaining work is increasing for '{ContainerName}'.", log.FormattedMessage);
-        }
-
-        [Fact]
-        public void GetScaleStatus_RemainingWorkDecreasing_ReturnsVote_ScaleIn()
-        {
-            var context = new ScaleStatusContext<CosmosDBTriggerMetrics>
-            {
-                WorkerCount = 1
-            };
-            var timestamp = DateTime.UtcNow;
-            context.Metrics = new List<CosmosDBTriggerMetrics>
-            {
-                new CosmosDBTriggerMetrics { RemainingWork = 150, PartitionCount = 0, Timestamp = timestamp.AddSeconds(15) },
-                new CosmosDBTriggerMetrics { RemainingWork = 100, PartitionCount = 0, Timestamp = timestamp.AddSeconds(15) },
-                new CosmosDBTriggerMetrics { RemainingWork = 80, PartitionCount = 0, Timestamp = timestamp.AddSeconds(15) },
-                new CosmosDBTriggerMetrics { RemainingWork = 40, PartitionCount = 0, Timestamp = timestamp.AddSeconds(15) },
-                new CosmosDBTriggerMetrics { RemainingWork = 20, PartitionCount = 0, Timestamp = timestamp.AddSeconds(15) },
-                new CosmosDBTriggerMetrics { RemainingWork = 10, PartitionCount = 0, Timestamp = timestamp.AddSeconds(15) },
-            };
-
-            var status = ((CosmosDBScaleMonitor)_listener.GetMonitor()).GetScaleStatus(context);
-            Assert.Equal(ScaleVote.ScaleIn, status.Vote);
-
-            var logs = _loggerProvider.GetAllLogMessages().ToArray();
-            var log = logs[0];
-            Assert.Equal(Microsoft.Extensions.Logging.LogLevel.Information, log.Level);
-            Assert.Equal($"Remaining work is decreasing for '{ContainerName}'.", log.FormattedMessage);
-        }
-
-        [Fact]
         public async Task StartAsync_Retries()
         {
             var attribute = new CosmosDBTriggerAttribute("test", "test") { LeaseContainerPrefix = ProcessorName };
